requirements.txt
absl-py==0.11.0
addict==2.4.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
colorama==0.4.4
ConfigArgParse==1.2.3
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.3
entrypoints==0.3
future==0.18.2
grpcio==1.34.0
imageio==2.4.0
imageio-ffmpeg==0.4.2
importlib-metadata==3.3.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.2.0
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
kornia==0.2.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.3.0
numpy==1.19.4
open3d==0.12.0
opencv-python==4.2.0.34
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plyfile==0.7.3
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.14.0
ptyprocess==0.7.0
pybind11==2.6.2
pycollada==0.7.1
pycparser==2.20
pyglet==1.5.16
Pygments==2.8.1
PyMCubes==0.1.2
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-lightning==0.7.5
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sourceinspect==0.0.4
taichi==0.7.16
taichi-glsl==0.0.10
taichi-three==0.0.9
tensorboard==1.15.0
tensorboardX==2.2
termcolor==1.1.0
terminado==0.9.4
test-tube==0.7.5
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.4.8
torch==1.8.1
torchsearchsorted @ file:///home/jcn/%E6%A1%8C%E9%9D%A2/Nerf/nerf_simple/torchsearchsorted
torchvision==0.9.1
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
trimesh==3.9.13
typing-extensions==3.7.4.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.0
